test: share HTTP stub for sync tests
